% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_mode.R
\name{assign_mode}
\alias{assign_mode}
\title{Assign most common user-provided value in case of discrepancy in patient
demographics information}
\usage{
assign_mode(multi_data_patients_tbl)
}
\arguments{
\item{multi_data_patients_tbl}{Tibble containing demographics and
sample_collection data for users with discrepancy in demographics
information}
}
\value{
Tibble containing demographics data for
users with discrepancy in demographics information replace by mode values
(most common value provided by the user)
}
\description{
Assign most common user-provided value in case of discrepancy in patient
demographics information
}
