% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_us_zip.R
\name{build_us_zip}
\alias{build_us_zip}
\title{Build a tibble with US Zip Code Information}
\usage{
build_us_zip(filepath)
}
\arguments{
\item{filepath}{Path to the csv file containing US zip codes
(downloaded from UnitedStatesZipCodes.org)}
}
\value{
Tibble containing US zip codes
}
\description{
Build a tibble with US Zip Code Information
}
