% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_biorepo.R
\name{compile_biorepo}
\alias{compile_biorepo}
\title{Compile Biorepository for inclusion in deidentifiedDB database}
\usage{
compile_biorepo(filepath)
}
\arguments{
\item{filepath}{Path to the biorepository csv file}
}
\value{
Tibble containing biorepository information for deidentifiedDB database
}
\description{
Compile Biorepository for inclusion in deidentifiedDB database
}
