% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_diagnostics_data.R
\name{compile_diagnostics_data}
\alias{compile_diagnostics_data}
\title{Compile weekly diagnostics data from REDDI before appending to the
deidentifiedDB SQLite database}
\usage{
compile_diagnostics_data(filepath, run_date_fmt = c("mdy"))
}
\arguments{
\item{filepath}{filepath to the diagnostics weekly data from REDDI}

\item{run_date_fmt}{Date format used to specify run date
(for example, 08-01-2020 for August 1, 2020)}
}
\value{
Tibble with diagnostics data from REDDI
}
\description{
Compile weekly diagnostics data from REDDI before appending to the
deidentifiedDB SQLite database
}
