% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_genbank.R
\name{compile_genbank}
\alias{compile_genbank}
\title{Compile Sequence and Metadata for GenBank Submission}
\usage{
compile_genbank(
  testkit_ids,
  seq_dir,
  sample_collection_tbl = NULL,
  demographics_tbl = NULL,
  viralrecon_tbl = NULL,
  deidentifiedDB = NULL
)
}
\arguments{
\item{testkit_ids}{Testkit IDs to include in the GenBank Submission}

\item{seq_dir}{Directory containing FASTA files of consensus sequence
for individual samples}

\item{sample_collection_tbl}{Path to the tibble containing
sample collection data}

\item{demographics_tbl}{Path to the tibble containing
demographics data}

\item{viralrecon_tbl}{Path to the tibble containing
viralrecon results}

\item{deidentifiedDB}{Path to the deidentifiedDB SQLite File}
}
\value{
List of three elements: 'int_tbl' : Tibble for internal use,
'ext_tbl' : Tibble containing metadata  required by GenBank,
'seqs' : DNAStringSet with consensus sequences for each Sequence_ID
}
\description{
Compile Sequence and Metadata for GenBank Submission
}
