% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_genbank_table.R
\name{compile_genbank_table}
\alias{compile_genbank_table}
\title{Compile GenBank Submission Information for Appending to deidentifiedDB}
\usage{
compile_genbank_table(
  accession_report_filpath,
  int_tbl,
  submission_date,
  pipeline_version,
  pipeline_name = "nf-core/viralrecon"
)
}
\arguments{
\item{accession_report_filpath}{Path to the Accession Report from GenBank}

\item{int_tbl}{'int_tbl' returned by compile_genbank function}

\item{submission_date}{Date of GenBank Submission}

\item{pipeline_version}{Version of pipeline that generated the
consensus sequence}

\item{pipeline_name}{Name of the pipeline that generated the
consensus sequence}
}
\value{
Tibble ready for integration into deidentifiedDB database
}
\description{
Compile GenBank Submission Information for Appending to deidentifiedDB
}
