% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_sc_data.R
\name{compile_sc_data}
\alias{compile_sc_data}
\title{Compile Sample Collection Data for deidentifiedDB SQLite Database}
\usage{
compile_sc_data(sc_tbl, us_zip_codes_filepath, global_regions_filepath)
}
\arguments{
\item{sc_tbl}{Tibble containing sample collection data}

\item{us_zip_codes_filepath}{Path to the csv file containing US zip codes
(downloaded from UnitedStatesZipCodes.org)}

\item{global_regions_filepath}{Path to the csv file of the UN Dataset downloaded
from https://unece.org/trade/cefact/UNLOCODE-Download)}
}
\value{
Final tibble containing sample collection data ready to be appended
to deidentifiedDB SQLite database
}
\description{
Compile Sample Collection Data for deidentifiedDB SQLite Database
}
