% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_viralrecon.R
\name{compile_viralrecon}
\alias{compile_viralrecon}
\title{Compile Results from nf-core/viralrecon for
inclusion in deidentifiedDB database}
\usage{
compile_viralrecon(
  filepath,
  run_date,
  viralrecon_version,
  primer_set_version,
  primer_set = "artic",
  variant_caller = "ivar",
  sequencing_platform = "illumina"
)
}
\arguments{
\item{filepath}{Path to the multiqc summary csv for the viralrecon run}

\item{run_date}{Start Date  of the viralrecon run
(for example, 2021-08-01 for August 1, 2021)}

\item{viralrecon_version}{Version of viralrecon used for this run}

\item{primer_set_version}{Primer set version (eg. 3, 4, 4.1)}

\item{primer_set}{Name of the primer set used (eg. 'artic' , 'midnight')}

\item{variant_caller}{Was 'iVar' or 'BCFTools' used for variant calling ?}

\item{sequencing_platform}{Sequencing Technology Used (eg. 'illumina',
'nanopore')}
}
\value{
Tibble containing results from nf-core/viralrecon for
inclusion to deidentifiedDB database
}
\description{
Compile Results from nf-core/viralrecon for
inclusion in deidentifiedDB database
}
