% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ct.R
\name{compute_ct}
\alias{compute_ct}
\title{Calculate Mean and Median Ct values}
\usage{
compute_ct(ct_tbl)
}
\arguments{
\item{ct_tbl}{Tibble with "testkit_id", "ct_rep1", and "ct_rep2" columns}
}
\value{
Tibble with mean and median Ct value for each \code{testkit_id}
}
\description{
Calculate Mean and Median Ct values
}
