% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fasta_genbank.R
\name{create_fasta_genbank}
\alias{create_fasta_genbank}
\title{Create DNAStringSet containing consensus sequences}
\usage{
create_fasta_genbank(input_dir, internal_metadata_tbl)
}
\arguments{
\item{input_dir}{Directory containing FASTA files of consensus sequence
for individual samples}

\item{internal_metadata_tbl}{int_tbl returned by get_metadata_genbank()}
}
\value{
DNAStringSet containing consensus sequences
}
\description{
Create DNAStringSet containing consensus sequences
}
