% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_control.R
\name{detect_control}
\alias{detect_control}
\title{Label samples on whether they were used as controls}
\usage{
detect_control(testkit_id, result)
}
\arguments{
\item{testkit_id}{Testkit ID}

\item{result}{Value from the 'result' column of the diagnostics table}
}
\value{
A logical value : TRUE if sample was used as a control, else FALSE
}
\description{
Label samples on whether they were used as controls
}
