% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_prelim_tidy_sc.R
\name{do_prelim_tidy_sc}
\alias{do_prelim_tidy_sc}
\title{Preliminary Tidy Up of the Demographics and Sample Collection Data Table}
\usage{
do_prelim_tidy_sc(
  sample_collection_tbl,
  date_fmt = c("\%m/\%d/\%y"),
  time_zone = "America/New_York"
)
}
\arguments{
\item{sample_collection_tbl}{Tibble obtained from read_demographics_csv()}

\item{date_fmt}{Format used to specify dates (Default: MM/DD/YYYY)}

\item{time_zone}{Time zone for collection time (Default: "America/New_York")}
}
\value{
Tibble with the Sample Collection and
the deidentified Demographics Data organized to make it convenient to update
deidentified DB database
}
\description{
Preliminary Tidy Up of the Demographics and Sample Collection Data Table
}
