% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daily_diagnostics.R
\name{get_daily_diagnostics}
\alias{get_daily_diagnostics}
\title{Title}
\usage{
get_daily_diagnostics(
  diagnostics_tbl,
  start_date,
  end_date,
  add_variables = NULL
)
}
\arguments{
\item{add_variables}{}
}
\value{

}
\description{
Title
}
