% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_global_entities.R
\name{get_global_entities}
\alias{get_global_entities}
\title{Build a Vector containing names of US and International states/provinces}
\usage{
get_global_entities(filepath)
}
\arguments{
\item{filepath}{Path to the csv file of the UN Dataset downloaded
from https://unece.org/trade/cefact/UNLOCODE-Download)}
}
\value{
Vector containing names of US and International states/provinces
}
\description{
Build a Vector containing names of US and International states/provinces
}
