% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata_genbank.R
\name{get_metadata_genbank}
\alias{get_metadata_genbank}
\title{Compile Metadata for GenBank Submission}
\usage{
get_metadata_genbank(
  testkit_ids,
  out_dir,
  sample_collection_tbl = NULL,
  demographics_tbl = NULL,
  viralrecon_tbl = NULL,
  deidentifiedDB = NULL
)
}
\arguments{
\item{testkit_ids}{Testkit IDs to include in the GenBank Submission}

\item{out_dir}{Path to the Output Directory}

\item{sample_collection_tbl}{Path to the tibble containing
sample collection data}

\item{demographics_tbl}{Path to the tibble containing
demographics data}

\item{viralrecon_tbl}{Path to the tibble containing
viralrecon results}

\item{deidentifiedDB}{Path to the deidentifiedDB SQLite File}
}
\value{
Tibble containing metadata information for genbank submission.
A tsv file is also written to out_dir
}
\description{
Compile Metadata for GenBank Submission
}
