% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nextclade_distribution.R
\name{get_nextclade_distribution}
\alias{get_nextclade_distribution}
\title{Get Nextclade Distribution of Sequenced Samples}
\usage{
get_nextclade_distribution(
  viralrecon_tbl,
  sample_collection_tbl,
  start_date = "2021-01-01",
  end_date = lubridate::date(lubridate::now()),
  n_days = 30
)
}
\arguments{
\item{viralrecon_tbl}{viralrecon table from deidentifiedDB database}

\item{sample_collection_tbl}{Sample Collection table from deidentifiedDB database}

\item{start_date}{Filter sample collection table for samples collected after
this date. Provide date as in "2021-08-01" for August 1, 2021}

\item{end_date}{Filter sample collection table for samples collected on or
before this date. Provide date as in "2021-08-01" for August 1, 2021}

\item{n_days}{Duration of time (in days) for which only the first collected
COVID-19 positive sample  for a patient is retained}
}
\value{
Tibble containing monthly count of sequenced samples in
each Nextclade
}
\description{
Get Nextclade Distribution of Sequenced Samples
}
