% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_redundant_samples.R
\name{get_redundant_samples}
\alias{get_redundant_samples}
\title{Get Testkit IDs of redundant COVID-19 positive samples {non-first (based on
collection date) COVID19 samples from probably a single-infection event}}
\usage{
get_redundant_samples(sc_tbl_no_missing, n_days = 30)
}
\arguments{
\item{sc_tbl_no_missing}{Sample collection tibble without missing patient ID,
Rymedi results, and collection date data}

\item{n_days}{Duration of time (in days) for which only the first collected
COVID-19 positive sample  for a patient is retained}
}
\value{
Vector containing Testkit IDs of redundant samples
}
\description{
Get Testkit IDs of redundant COVID-19 positive samples {non-first (based on
collection date) COVID19 samples from probably a single-infection event}
}
