% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_us_entities.R
\name{get_us_entities}
\alias{get_us_entities}
\title{Build a Vector containing US State Codes}
\usage{
get_us_entities(filepath)
}
\arguments{
\item{filepath}{Path to the csv file containing US zip codes
(downloaded from UnitedStatesZipCodes.org)}
}
\value{
Vector containing two letter codes of US states and territories
}
\description{
Build a Vector containing US State Codes
}
