% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_patient.R
\name{inspect_patient}
\alias{inspect_patient}
\title{Inspect deidentified demographics data for each patient}
\usage{
inspect_patient(demographics_tbl, poi)
}
\arguments{
\item{demographics_tbl}{deidentified demographics data}

\item{poi}{hashed patient ID from REDDI}
}
\value{
deidentified demographics data for each patient
}
\description{
Inspect deidentified demographics data for each patient
}
