% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_zip_code.R
\name{lookup_zip_code}
\alias{lookup_zip_code}
\title{Compile Reference Location Information}
\usage{
lookup_zip_code(location_tbl, us_zip_codes_filepath, global_regions_filepath)
}
\arguments{
\item{location_tbl}{Tibble with "zip_code", "state", and "city" columns}

\item{us_zip_codes_filepath}{Path to the csv file containing US zip codes
(downloaded from UnitedStatesZipCodes.org)}

\item{global_regions_filepath}{Path to the csv file of the UN Dataset downloaded
from https://unece.org/trade/cefact/UNLOCODE-Download)}
}
\value{
Tibble with validated USPS location information
}
\description{
Compile Reference Location Information
}
