% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_sc.R
\name{polish_sc}
\alias{polish_sc}
\title{Polish Sample Collection Tibble}
\usage{
polish_sc(sc_tbl)
}
\arguments{
\item{sc_tbl}{Tibble containing sample collection data
after tidy_up_location()}
}
\value{
Final tibble containing sample collection data ready to be appended
to deidentifiedDB SQLite database
}
\description{
Polish Sample Collection Tibble
}
