% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_demographics_sc.R
\name{prepare_demographics_sc}
\alias{prepare_demographics_sc}
\title{Transform COVID19 Sample Collection and deidentified Demographics Data
from REDDI for the deidentifiedDB demographics and sample_collection tables}
\usage{
prepare_demographics_sc(
  filepath,
  date_fmt = c("\%m/\%d/\%y"),
  time_zone = "America/New_York"
)
}
\arguments{
\item{filepath}{Path to the CSV file containing Sample Collection and
deidentified Demographics Data}

\item{date_fmt}{Format used to specify dates (Default: MM/DD/YYYY)}

\item{time_zone}{Time zone for collection time (Default: "America/New_York")}
}
\value{
Tibble with the Sample Collection and
the deidentified Demographics Data organized to make it convenient to update
deidentified DB database
}
\description{
Transform COVID19 Sample Collection and deidentified Demographics Data
from REDDI for the deidentifiedDB demographics and sample_collection tables
}
