% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_demographics.R
\name{pull_demographics}
\alias{pull_demographics}
\title{Extract demographics data from table returned by prepare_demographics_sc}
\usage{
pull_demographics(demographics_sc_tbl)
}
\arguments{
\item{demographics_sc_tbl}{Tibble returned by the
prepare_demographics_sc function}
}
\value{
Tibble with only demographics data
}
\description{
Extract demographics data from table returned by prepare_demographics_sc
}
