% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_sc.R
\name{pull_sc}
\alias{pull_sc}
\title{Extract sample collection data from table returned by prepare_demographics_sc}
\usage{
pull_sc(demographics_sc_tbl)
}
\arguments{
\item{demographics_sc_tbl}{Tibble containing sample collection
and demographics information from REDDIT}
}
\value{
Tibble with only sample collection data
}
\description{
Extract sample collection data from table returned by prepare_demographics_sc
}
