% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_demographics_csv.R
\name{read_demographics_csv}
\alias{read_demographics_csv}
\title{Read CSV file containing COVID19 Sample Collection and
deidentified Demographics Data from REDDI}
\usage{
read_demographics_csv(
  filepath,
  date_fmt = c("\%m/\%d/\%y"),
  time_zone = "America/New_York"
)
}
\arguments{
\item{filepath}{Path to the CSV file containing Sample Collection and
deidentified Demographics Data}

\item{date_fmt}{Format used to specify dates (Default: MM/DD/YYYY)}

\item{time_zone}{Time zone for collection time (Default: "America/New_York")}
}
\value{
A tibble with the Sample Collection and
the deidentified Demographics Data
}
\description{
Read CSV file containing COVID19 Sample Collection and
deidentified Demographics Data from REDDI
}
