% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_diagnostics_csv.R
\name{read_diagnostics_csv}
\alias{read_diagnostics_csv}
\title{Read CSV file containing COVID19 Diagnostics Data from REDDI}
\usage{
read_diagnostics_csv(filepath)
}
\arguments{
\item{filepath}{Path to the CSV file containing COVID19 Diagnostics Data}
}
\value{
A tibble with the COVID19 Diagnostics Data under column names used in the deidentifiedDB SQLite database.
}
\description{
Read CSV file containing COVID19 Diagnostics Data from REDDI
}
