% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_up_location.R
\name{tidy_up_location}
\alias{tidy_up_location}
\title{Tidy Up Location Information}
\usage{
tidy_up_location(sc_tbl, location_combinations)
}
\arguments{
\item{sc_tbl}{Tibble containing sample collection data}

\item{location_combinations}{Tibble with validated USPS location information
from lookup_zip_code()}
}
\value{
Tidy Tibble with Location Information
}
\description{
Tidy Up Location Information
}
