% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_up_race.R
\name{tidy_up_race}
\alias{tidy_up_race}
\title{Tidy up the \code{race} column}
\usage{
tidy_up_race(demographics_tbl)
}
\arguments{
\item{demographics_tbl}{Tibble with demographics data}
}
\value{
Tibble with demographics data with race information tidied up
}
\description{
Tidy up the \code{race} column
}
