% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_up_zip.R
\name{tidy_up_zip}
\alias{tidy_up_zip}
\title{Remove non-numeric characters from zip codes and extract 5 digit zip code}
\usage{
tidy_up_zip(zip_code)
}
\arguments{
\item{zip_code}{input zip code}
}
\value{
5 character long zip code
}
\description{
Remove non-numeric characters from zip codes and extract 5 digit zip code
}
