% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daily_diagnostics.R
\name{get_daily_diagnostics}
\alias{get_daily_diagnostics}
\title{Get Daily Diangostics Counts}
\usage{
get_daily_diagnostics(
  diagnostics_tbl,
  start_date,
  end_date,
  add_variables = NULL
)
}
\arguments{
\item{diagnostics_tbl}{Tibble containing the following columns 1)testkit_id,
2) run_Date
3)result}

\item{start_date}{Filter sample collection table for samples collected after
this date. Provide date as in "2021-08-01" for August 1, 2021}

\item{end_date}{Filter sample collection table for samples collected on or
before this date. Provide date as in "2021-08-01" for August 1, 2021}

\item{add_variables}{variables to group_by}
}
\value{
Tibble with daily counts of different diagnostics result categories
}
\description{
Get Daily Diangostics Counts
}
