% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata_genbank.R
\name{get_metadata_genbank}
\alias{get_metadata_genbank}
\title{Compile Metadata for GenBank Submission}
\usage{
get_metadata_genbank(
  testkit_ids,
  sample_collection_tbl = NULL,
  demographics_tbl = NULL,
  viralrecon_tbl = NULL,
  deidentifiedDB = NULL,
  time_zone = "America/New_York"
)
}
\arguments{
\item{testkit_ids}{Testkit IDs to include in the GenBank Submission}

\item{sample_collection_tbl}{Path to the tibble containing
sample collection data}

\item{demographics_tbl}{Path to the tibble containing
demographics data}

\item{viralrecon_tbl}{Path to the tibble containing
viralrecon results}

\item{deidentifiedDB}{Path to the deidentifiedDB SQLite File}

\item{time_zone}{Time zone for collection time (Default: "America/New_York")}
}
\value{
Tibble containing metadata information for genbank submission.
}
\description{
Compile Metadata for GenBank Submission
}
