% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weekly_tpr.R
\name{plot_weekly_tpr}
\alias{plot_weekly_tpr}
\title{Plot Weekly Test Positivity Rate}
\usage{
plot_weekly_tpr(diagnostics_tbl, start_date = NULL, end_date = NULL)
}
\arguments{
\item{diagnostics_tbl}{Tibble with diagnostics data (should contain at least
two columns: 'run date' and 'result')}

\item{start_date}{start date for the plot}

\item{end_date}{end date for the plot}
}
\value{

}
\description{
Plot Weekly Test Positivity Rate
}
