% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_demographics_csv.R
\name{read_demographics_csv}
\alias{read_demographics_csv}
\title{Read CSV file containing COVID19 Sample Collection and
deidentified Demographics Data from REDDI}
\usage{
read_demographics_csv(filepath)
}
\arguments{
\item{filepath}{Path to the CSV file containing Sample Collection and
deidentified Demographics Data}
}
\value{
A tibble with the Sample Collection and
the deidentified Demographics Data
}
\description{
Read CSV file containing COVID19 Sample Collection and
deidentified Demographics Data from REDDI
}
